refactor(metrics): preserve module headroom
