+++ scripts/gate/routing-table/arms-tooling-modules.mjs
@@ -56,7 +56,9 @@ export const TOOLING_MODULE_ARMS = [
       "scripts/pr/review-process-metrics.mjs",
       "scripts/pr/review-process-metrics-actions.mjs",
       "scripts/pr/review-process-metrics-core.mjs",
+      "scripts/pr/review-process-metrics-evidence.mjs",
       "scripts/pr/review-process-metrics-finding-classifier.mjs",
+      "scripts/pr/review-process-metrics-finding-language.mjs",
       "scripts/pr/review-process-metrics-finding-preflight.mjs",
       "scripts/pr/review-process-metrics-legacy.mjs",
       "scripts/pr/review-process-metrics-markdown.mjs",
